+++ packages/web-crawler/src/urlRules.ts
@@ -22,6 +22,11 @@ export const crawUrlRules: CrawlUrlRule[] = [
     impls: ['jina'],
     urlPattern: 'https://(.*).pdf',
   },
+  // arxiv PDF use jina
+  {
+    impls: ['jina'],
+    urlPattern: 'https://arxiv.org/pdf/(.*)',
+  },
   // 知乎有爬虫防护，使用 jina
   {
     impls: ['jina'],
@@ -32,7 +37,13 @@ export const crawUrlRules: CrawlUrlRule[] = [
     urlPattern: 'https://medium.com/(.*)',
     urlTransform: 'https://scribe.rip/$1',
   },
-
+  {
+    filterOptions: {
+      enableReadability: false,
+    },
+    impls: ['jina', 'browserless'],
+    urlPattern: 'https://(twitter.com|x.com)/(.*)',
+  },
   // 体育数据网站规则
   {
     filterOptions: {